requirements.txt
# Core dependencies for ChatPDF clone
google-generativeai>=0.3.0
PyPDF2>=3.0.1
pdfplumber>=0.9.0
click>=8.1.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
